Batch-update finished emails count during sending
When we have email campaigns with many emails, the campaign row will be
updated each time an email is sent/failed which can produce lots of
“dead” rows and unnecessary updates. As this information is used just
for stats and is not time-critical, we can safely update the counters
once for each batch sending which will greatly reduce the number of SQL
update statements.

The second issue this commit solves is that concurrent updates will now
be safe as opposed to before where we did set the mails count to a
fixed number.