Add creating of QueuedMail with custom template
Renames model MailCampaignAttachment to MailAttachment

Adds model MailTemplate which has MailAttachment-s

MailCampaign has MailTemplate
QueuedMail could also have a MailTemplate

When you make a change like QueuedMail.body_html = '<h1>test</h1>'
it will copy the MailTemplate from the MailCampaign and then it will
update the copy with your changes.

MailCampaign#add_attachment does not save the attachment you have to
call MailCampaign#save, the same is for QueuedMail

Usually you will work with MailCampaign or QueuedMail and you won't
touch the MailTemplate and MailAttachment

Adds QueuedMail#require_uniqueness which is a boolean and it is either
true or nil. This allows the unique index on [:to, :mail_campaign, :require_uniqueness]
in the database to work as expected (a.k.a. if nil then not unique).