Fix unique message key initialization
A unique per-email key is generated in QueuedMail. This key should be
preserved once generated as it is used for interpolation in the email
body and then copied to the FinishedMail so that the sent email can be
tracked and so that unsubscribe works.

We broke this when we added a random component in the message key in
6ec62ea7. It has been broken since June 26th, v0.8.0.

The change now keeps the value of the message key, once generated. Also,
QueuedMail#id no longer is included in the key as it will never be
present there as the key is also generated before validation, when we
don’t have an id yet. The secure random component should be enough for
uniqueness.